Implement Exponential Moving Average (EMA) for affinity
- Models: Update `ResonanceEdge.affinity_level` to `float` (default 50.0).
- Local Repo: Implement EMA logic in `consolidate_memories` with Alpha=0.15.
- Local Repo: Update `update_resonance` to handle float deltas.
- Spanner Repo: Update `update_resonance` to use `FLOAT64` and handle float affinity.
- Interfaces: Update `SoulRepository` abstract method signature.
- Logic: Map average valence (-1.0 to 1.0) to target signal (0-100) before archiving episodes.

+++ backend/app/repositories/base.py
@@ -32,7 +32,7 @@ class SoulRepository(ABC):
         pass
 
     @abstractmethod
-    async def update_resonance(self, user_id: str, agent_id: str, delta: int) -> ResonanceEdge:
+    async def update_resonance(self, user_id: str, agent_id: str, delta: float) -> ResonanceEdge:
         """Update the affinity level of a resonance edge."""
         pass
